// src/layers/connection-layer/types/request-events.enum.ts
export enum RequestEventsEnum {
  CALL_ACTION = 'call_action',
  REPLY_ACTION = 'reply_action',
  SEND_MESSAGE = 'send_message',
  CREATE_CHANNEL = 'create_channel',
  SEND_TO_CHANNEL = 'send_to_channel',
  JOIN_CONNECTION_TO_CHANNELS = 'join_connection_to_channels',
  LEAVE_CONNECTION_TO_CHANNELS = 'leave_connection_to_channels',
}